refactor: consolidate LLM model definitions into domain layer
- Move MODELS_CONFIG and DEFAULT_MODEL from backend/configs/ to backend/domain/models/llm_model.py
- Remove backend/configs/ directory entirely
- Update all imports across the codebase to reference domain layer directly
- Update architecture documentation to reflect the simplified structure

Benefits:
- Eliminates confusion between config.py (environment variables) and configs/ (static data)
- Consolidates model entity definition and model catalog in one place
- Clearer separation: config.py for runtime configuration, domain models for business data
- Follows DDD principle of keeping domain data with domain entities
