--- backend/configs/__init__.py
@@ -1,4 +1,0 @@
-"""Configuration package.
-
-設定データを管理します。
-"""
