+++ backend/domain/models/llm_model.py
@@ -1,6 +1,6 @@
 """LLM モデルのドメインモデル。
 
-LLM モデルの情報を表現するエンティティです。
+LLM モデルの情報を表現するエンティティと、利用可能なモデルの定義を管理します。
 """
 
 from pydantic import BaseModel
@@ -24,3 +24,42 @@ class LLMModel(BaseModel):
     input_cost_per_1m: float
     output_cost_per_1m: float
     is_free: bool
+
+
+# 利用可能な LLM モデルの定義
+MODELS_CONFIG: dict[str, LLMModel] = {
+    "xiaomi/mimo-v2-flash:free": LLMModel(
+        id="xiaomi/mimo-v2-flash:free",
+        name="MIMO v2 Flash",
+        provider="openrouter",
+        input_cost_per_1m=0.0,
+        output_cost_per_1m=0.0,
+        is_free=True,
+    ),
+    "mistralai/devstral-2512:free": LLMModel(
+        id="mistralai/devstral-2512:free",
+        name="DevStral-2512",
+        provider="openrouter",
+        input_cost_per_1m=0.0,
+        output_cost_per_1m=0.0,
+        is_free=True,
+    ),
+    "x-ai/grok-4.1-fast": LLMModel(
+        id="x-ai/grok-4.1-fast",
+        name="Grok 4.1 Fast",
+        provider="openrouter",
+        input_cost_per_1m=0.20,
+        output_cost_per_1m=0.50,
+        is_free=False,
+    ),
+    "deepseek/deepseek-v3.2": LLMModel(
+        id="deepseek/deepseek-v3.2",
+        name="DeepSeek v3.2",
+        provider="openrouter",
+        input_cost_per_1m=0.25,
+        output_cost_per_1m=0.38,
+        is_free=False,
+    ),
+}
+
+DEFAULT_MODEL = "xiaomi/mimo-v2-flash:free"
