+++ backend/tests/unit/tools/spotify/test_stats.py
@@ -75,9 +75,7 @@ class TestGetTopTracksTool:
         tool = GetTopTracksTool(mock_repository)
 
         # Act: ツールを実行
-        result = tool.execute(
-            start_date="2024-01-01", end_date="2024-01-31", limit=10
-        )
+        result = tool.execute(start_date="2024-01-01", end_date="2024-01-31", limit=10)
 
         # Assert: 実行結果とリポジトリ呼び出しを検証
         assert len(result) == 1
